# Rewrite the grammar once again