--- a/.gitattributes
+++ b/.gitattributes
@@ -1,2 +1,5 @@
 /src/** linguist-vendored
-/examples/* linguist-vendored
+/test/libs/* linguist-vendored
+/src/parser.c -diff
+/src/grammar.json -diff
+/src/node-types.json -diff