--- a/.npmignore
+++ b/.npmignore
@@ -1,6 +1,0 @@
-corpus
-examples
-build
-script
-.travis.yml
-.gitattributes