--- a/binding.gyp
+++ b/binding.gyp
@@ -2,19 +2,21 @@
   "targets": [
     {
       "target_name": "tree_sitter_haskell_binding",
+      "dependencies": [
+        "<!(node -p \"require('node-addon-api').targets\"):node_addon_api_except",
+      ],
       "include_dirs": [
-        "<!(node -e \"require('nan')\")",
-        "src"
+        "src",
       ],
       "sources": [
-        "src/parser.c",
         "bindings/node/binding.cc",
+        "src/parser.c",
         "src/scanner.c",
         "src/unicode.h",
       ],
       "cflags_c": [
-        "-std=c99",
-      ]
+        "-std=c11",
+      ],
     }
   ]
 }