--- a/bindings/c/tree-sitter-haskell.h
+++ b/bindings/c/tree-sitter-haskell.h
@@ -1,0 +1,16 @@
+#ifndef TREE_SITTER_HASKELL_H_
+#define TREE_SITTER_HASKELL_H_
+
+typedef struct TSLanguage TSLanguage;
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+const TSLanguage *tree_sitter_haskell(void);
+
+#ifdef __cplusplus
+}
+#endif
+
+#endif // TREE_SITTER_HASKELL_H_