--- a/bindings/c/tree-sitter-haskell.pc.in
+++ b/bindings/c/tree-sitter-haskell.pc.in
@@ -1,0 +1,11 @@
+prefix=@PREFIX@
+libdir=@LIBDIR@
+includedir=@INCLUDEDIR@
+
+Name: tree-sitter-haskell
+Description: Haskell grammar for tree-sitter
+URL: @URL@
+Version: @VERSION@
+Requires: @REQUIRES@
+Libs: -L${libdir} @ADDITIONAL_LIBS@ -ltree-sitter-haskell
+Cflags: -I${includedir}