--- a/bindings/go/binding.go
+++ b/bindings/go/binding.go
@@ -1,0 +1,13 @@
+package tree_sitter_haskell
+
+// #cgo CFLAGS: -std=c11 -fPIC
+// #include "../../src/parser.c"
+// // NOTE: if your language has an external scanner, add it here.
+import "C"
+
+import "unsafe"
+
+// Get the tree-sitter Language for this grammar.
+func Language() unsafe.Pointer {
+	return unsafe.Pointer(C.tree_sitter_haskell())
+}