--- a/bindings/go/go.mod
+++ b/bindings/go/go.mod
@@ -1,0 +1,5 @@
+module github.com/tree-sitter/tree-sitter-haskell
+
+go 1.22
+
+require github.com/smacker/go-tree-sitter v0.0.0-20230720070738-0d0a9f78d8f8