--- a/bindings/python/tree_sitter_haskell/__init__.py
+++ b/bindings/python/tree_sitter_haskell/__init__.py
@@ -1,0 +1,5 @@
+"Haskell grammar for tree-sitter"
+
+from ._binding import language
+
+__all__ = ["language"]