--- a/cabal.project
+++ b/cabal.project
@@ -1,0 +1,1 @@
+packages: tools