--- a/queries/locals.scm
+++ b/queries/locals.scm
@@ -1,4 +1,4 @@
 (signature name: (variable)) @local.definition
 (function name: (variable)) @local.definition
-(pat_name (variable)) @local.definition
-(exp_name (variable)) @local.reference
+(pattern/variable) @local.definition
+(expression/variable) @local.reference