--- a/tools/app/Main.hs
+++ b/tools/app/Main.hs
@@ -1,0 +1,3 @@
+module Main (main) where
+
+import TSH (main)