--- a/tools/lib/TSH/Monad.hs
+++ b/tools/lib/TSH/Monad.hs
@@ -1,0 +1,2 @@
+module TSH.Monad where
+