… .github/workflows/assets.yml
… .github/workflows/ci.yml
… .github/workflows/legacy.yml
… .github/workflows/release.yml
… .gitignore
… Cargo.lock
… Cargo.toml
… README.md
… bindings/go/binding_test.go
… bindings/python/tree_sitter_haskell/__init__.pyi
… flake.lock
… grammar/basic.js
… grammar/class.js
… grammar/conflicts.js
… grammar/data.js
… grammar/decl.js
… grammar/exp.js
… grammar/externals.js
… grammar/general.js
… grammar/id.js
… grammar/import.js
… grammar/inline.js
… grammar/lexeme.js
… grammar/literal.js
… grammar/misc.js
… grammar/module.js
… grammar/operator.js
… grammar/pat.js
… grammar/patsyn.js
… grammar/pattern.js