@clason 2024-03-24
> I would also prefer to change the workflow so that parser.c is only generated on a release branch, not for
every commit, like some other grammars do it.

In this case, please at least commit the `grammar.json`, which is much smaller and has meaningful diffs.

Important question: will queries remain compatible?