@tek 2024-03-24
> In this case, please at least commit the `grammar.json`, which is much smaller and has meaningful diffs. This will allow people to generate the parser without having to use node.

Ah, didn't know that. How do you generate the parser from that file?

> You can use `tree-sitter build --wasm` with the latest version. Please upload that as a release artifact as well.

That's how we're building it, but @wenkokke added a patch for `web-tree-sitter` for some reason I don't recall.

I regenerated the `.wasm` file and included it in the PR. Are you talking about Github release artifacts?

> Important question: will queries remain compatible?

Some of them will, but it's gonna depend on the result of the survey how contained the breakage will be.
I think most expression queries will still work, but types and top level structures should be affected.