@clason 2024-03-24
> Ah, didn't know that. How do you generate the parser from that file?

Just `tree-sitter generate grammar.json` (little known fact).

> That's how we're building it, but @wenkokke added a patch for web-tree-sitter for some reason I don't recall.

Possibly because the output dir has changed? But that should be fixed upstream.

> I regenerated the .wasm file and included it in the PR. Are you talking about Github release artifacts?

Yes, the release artifacts (only; it shouldn't be committed to the repo).

> Some of them will, but it's gonna depend on the result of the survey how contained the breakage will be.
I think most expression queries will still work, but types and top level structures should be affected.

Thanks for the heads-up. I hope the queries here can serve as a guide for nvim-treesitter.