@wenkokke 2024-03-24
> That's how we're building it, but @wenkokke added a patch for `web-tree-sitter` for some reason I don't recall.

The reason is that web-tree-sitter doesn't support all of libc, and there's a file in the web-tree-sitter source that determines exactly what symbols are supported.

Please test the WASM build before merging. Tests using only the standard testing infrastructure sadly aren't reliable.