@tek 2024-03-24
@414owen thanks so much! I definitely went a bit overboard with the obsession over the last few weeks 😁 
I don't think it's necessary to review the scanner in detail – running that fuzzer should be enough to complement the tests.