@mrcjkb 2024-03-25
> Important question: will queries remain compatible?

@clason Since I've worked on a lot of the queries in nvim-treesitter, and also maintain [neotest-haskell](https://github.com/mrcjkb/neotest-haskell), which also uses tree-sitter queries, I'll be taking part in the survey.
Personally, I'm quite excited about some of the proposed changes and wouldn't mind backward compatibility breaking. And of course, I'd be happy to adjust the queries in nvim-treesitter.