@clason 2024-03-25
@mrcjkb Thank you, then I don't need to worry :) (To be clear, I didn't want to argue for _not_ changing the queries; I just wanted a heads-up how much breakage I should expect so I can brace for it.)