@tek 2024-03-25
@414owen I'll add comments to the scanner later, so you won't have to keep guessing 😄