@wenkokke 2024-04-02
> @amaanq can you tell me whether the scanner is ever executed concurrently? I've made the state pointer a global variable that I assign in `tree_sitter_haskell_external_scanner_scan` to avoid having it as a parameter of every function, so it's not thread safe.

Would this result in problems if a single instance of the library is used, e.g., for multiple tabs in VSCode?